+++ src/util/timestampHoverProvider.ts
@@ -17,7 +17,7 @@ class TimestampHoverProvider implements vscode.HoverProvider, vscode.Disposable 
 
     private _disposables: vscode.Disposable[];
 
-    private _targetFormat: string;
+    private _targetFormats: Array<string | { customFormat: string, name: string, localize: boolean }>;
     private _enabled: boolean;
 
     constructor(timeConverter: TimeConverter) {
@@ -44,26 +44,11 @@ class TimestampHoverProvider implements vscode.HoverProvider, vscode.Disposable 
             const hoveredWord = document.getText(timeRange);
             if (this._timeConverter.isValidEpoch(hoveredWord)) {
                 const input = new InputDefinition(hoveredWord);
-                const prefix = '*Epoch Unit*: `' + input.originalUnit + '`  \n*Timestamp*: `';
 
                 // Is the feature enabled?
                 if (this._enabled) {
-                    if (this._targetFormat === 'utc') {
-                        const utc = this._timeConverter.epochToISOUtc(input.inputAsMs.toString());
-                        result = new vscode.Hover(prefix + utc + '`', timeRange);
-                    } else if (this._targetFormat === 'local') {
-                        const local = this._timeConverter.epochToIsoLocal(input.inputAsMs.toString());
-                        result = new vscode.Hover(prefix + local + '`', timeRange);
-                    } else if (this._targetFormat) {
-                        const custom = this._timeConverter.epochToCustom(
-                            input.inputAsMs.toString(),
-                            this._targetFormat);
-                        result = new vscode.Hover(prefix + custom + '`', timeRange);
-                    } else {
-                        result = undefined;
-                    }
-                } else {
-                    result = undefined;
+                    const message = this.buildMessage(input);
+                    result = message ? new vscode.Hover(message) : undefined;
                 }
             }
         }
@@ -75,11 +60,55 @@ class TimestampHoverProvider implements vscode.HoverProvider, vscode.Disposable 
         this._disposables.forEach((disposable) => disposable.dispose());
     }
 
-    private updateTimestampTargetFormat(): void {
-        const config = vscode.workspace.getConfiguration('timing.hoverTimestamp')
-            .get<string>('targetFormat', 'utc');
+    private buildMessage(input: InputDefinition): string {
+        let result: string = '';
 
-        this._targetFormat = config;
+        // Append message foreach configured format
+        for (const format of this._targetFormats) {
+            if (format === 'utc') {
+                // UTC format
+                const utc = this._timeConverter.epochToISOUtc(input.inputAsMs.toString());
+                result += '  \n*UTC Timestamp*: `' + utc + '`';
+
+            } else if (format === 'local') {
+                // Local format
+                const local = this._timeConverter.epochToIsoLocal(input.inputAsMs.toString());
+                result += '  \n*Local Timestamp*: `' + local + '`';
+
+            } else if (typeof format === 'string') {
+                // Other simple custom format
+                if (format) {
+                    const custom = this._timeConverter.epochToCustom(
+                        input.inputAsMs.toString(),
+                        format);
+                    result += '  \n*Formatted Timestamp*: `' + custom + '`';
+                }
+
+            } else if (format.customFormat) {
+                // Advanced configured custom format
+                const custom = this._timeConverter.epochToCustom(
+                    input.inputAsMs.toString(),
+                    format.customFormat,
+                    format.localize);
+                if (format.name) {
+                    result += '  \n*' + format.name + '*: `' + custom + '`';
+                } else {
+                    result += '  \n*Formatted Timestamp*: `' + custom + '`';
+                }
+            }
+        }
+
+        return result ? '*Epoch Unit*: `' + input.originalUnit + '`' + result : undefined;
+    }
+
+    private updateTimestampTargetFormat(): void {
+        const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
+        const value = config.get<any>('targetFormat', ['utc']);
+        if (typeof value === 'string') {
+            this._targetFormats = [value];
+        } else if (value instanceof Array) {
+            this._targetFormats = value;
+        }
     }
 
     private updateTimestampEnabled(): void {