+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "timing" extension will be documented in this file. T
 ## Unreleased
 * **Changed** `timing.hoverTimestamp.targetFormat` setting. It now supports an array of multiple formats, too.  
   Further in the array you can additionally express custom formats as objects for advanced configuration.
+* **Added** webpack for bundling the extension. Now the extension ships with 11 files instead of 383. Therefore the extension now is loaded much faster.
 
 ## 2019-03-16 - v2.0.0
 * **BREAKING CHANGES**: